Removes previous impl and adds better documentation for new impl
